Moved tests and files inside lib folder
