+++ lib/__tests__/bing.js
@@ -1,10 +1,10 @@
-var subquest = require('../')
+var subquest = require('../index')
 
 describe('Usig bing domain dork queries', function() {
 
   describe('get the array of subdomains', function() {
     jest.setTimeout(10000)
-    
+
     // Test bing search against google.com
     it('google.com', (done) => {
       subquest.getSubDomains({
