+++ lib/__tests__/main.js
@@ -1,4 +1,4 @@
-var subquest = require('../')
+var subquest = require('../index')
 
 var options = {
   host: 'google.com'
