+++ lib/__tests__/resolvers.js
@@ -1,4 +1,4 @@
-var subquest = require('../index.js')
+var subquest = require('../index')
 
 describe('Get Resolvers:', function() {
