+++ lib/__tests__/server.js
@@ -1,4 +1,4 @@
-var subquest = require('../index.js')
+var subquest = require('../index')
 
 describe('Valid DNS Servers', function() {
