+++ lib/bingSearch.js
@@ -53,5 +53,4 @@ exports.find = function(domainName, callback = () => {}) {
     list = _.uniq(list);
     callback(null, list);
   })
-
 }
